LayoutTests/platform/gtk/accessibility/text-for-range-simple-expected.txt

This is a test.

Hello world.
This tests the ability to get element text for a specified range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS paragraph.stringForRange(0, -1) is 'This is a test.'
PASS block.stringForRange(0, 12) is 'Hello world.'
PASS successfullyParsed is true

TEST COMPLETE
